+++ src/core/GameFile.h
@@ -49,7 +49,7 @@ class _GAMEFILE_API_ GameFile : public Component
     QString fullname() const { return filepath; }
     int fileDataId() { return m_fileDataId; }
 
-    void allocate(unsigned int size);
+    void allocate(unsigned long long size);
     bool setChunk(std::string chunkName, bool resetToStart = true);
     bool isChunked() { return chunks.size() > 0; }
 
@@ -59,14 +59,14 @@ class _GAMEFILE_API_ GameFile : public Component
 
     virtual bool openFile() = 0;
     virtual bool isAlreadyOpened() = 0;
-    virtual bool getFileSize(unsigned int & s) = 0;
+    virtual bool getFileSize(unsigned long long & s) = 0;
     virtual unsigned long readFile() = 0;
     virtual void doPostOpenOperation() = 0;
     virtual bool doPostCloseOperation() = 0;
 
     bool eof;
     unsigned char *buffer;
-    unsigned int pointer, size;
+    unsigned long long pointer, size;
     QString filepath;
     int m_fileDataId;
