+++ src/games/wow/CASCFile.h
@@ -38,7 +38,7 @@ class _CASCFILE_API_ CASCFile : public GameFile
   protected:
     virtual bool openFile();
     virtual bool isAlreadyOpened();
-    virtual bool getFileSize(unsigned int & s);
+    virtual bool getFileSize(unsigned long long & s);
     virtual unsigned long readFile();
     virtual void doPostOpenOperation();
     virtual bool doPostCloseOperation();
